リファクタリング: コンポーネントを切り出し
